+++ lib/classes/EzStoredImage.dart
@@ -1,10 +1,16 @@
 library empathetech_flutter_ui;
 
+import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
+
+import 'dart:io';
 import 'package:flutter/material.dart';
 
-class EzImage extends Image {
+class EzStoredImage extends Image {
   final Key? key;
-  final ImageProvider<Object> image;
+
+  /// [EzConfig] key that contains the path to the image you wish to load
+  final String prefsKey;
+
   final Widget Function(BuildContext, Widget, int?, bool)? frameBuilder;
   final Widget Function(BuildContext, Widget, ImageChunkEvent?)? loadingBuilder;
   final Widget Function(BuildContext, Object, StackTrace?)? errorBuilder;
@@ -23,10 +29,12 @@ class EzImage extends Image {
   final bool isAntiAlias;
   final FilterQuality filterQuality;
 
-  /// [Image] that requires a [semanticLabel] for encforcing accessibility
-  EzImage({
+  /// [Image] wrapper for when [prefsKey] can resolve to either an [AssetImage] or [FileImage]
+  /// If the [ImageProvider] is known, it is preferred to called the standard const [Image] constructor
+  /// Also requires a [semanticLabel] for encforcing accessibility
+  EzStoredImage({
     this.key,
-    required this.image,
+    required this.prefsKey,
     this.frameBuilder,
     this.loadingBuilder,
     this.errorBuilder,
@@ -45,8 +53,8 @@ class EzImage extends Image {
     this.isAntiAlias = false,
     this.filterQuality = FilterQuality.low,
   }) : super(
+          image: getProvider(prefsKey),
           key: key,
-          image: image,
           frameBuilder: frameBuilder,
           loadingBuilder: loadingBuilder,
           errorBuilder: errorBuilder,
@@ -66,4 +74,34 @@ class EzImage extends Image {
           isAntiAlias: isAntiAlias,
           filterQuality: filterQuality,
         );
+}
+
+/// Automatically handles [AssetImage] vs [FileImage]
+/// Technically supports [NetworkImage], but at this time it isn't recommended
+/// Will use [EzConfig] preferences as a backup if the [EzConfig] prefs call fails
+/// In a total failure event, a stock owl image will be shown
+ImageProvider getProvider(String prefsKey) {
+  // Something went wrong, return watchful owl
+  const String errorURL =
+      'https://flutter.github.io/assets-for-api-docs/assets/widgets/owl-2.jpg';
+
+  dynamic prefsValue = EzConfig.instance.prefs[prefsKey];
+
+  final String path = (prefsValue is String)
+      ? prefsValue
+      : (EzConfig.instance.preferences.getString(prefsKey) ?? errorURL);
+
+  if (isAsset(path)) {
+    return AssetImage(path);
+  } else {
+    try {
+      return FileImage(File(path));
+    } on FileSystemException catch (_) {
+      // File not found error - wipe the setting and return the/a backup image
+      EzConfig.instance.preferences.remove(prefsKey);
+
+      // Default case, stock owl
+      return NetworkImage(path);
+    }
+  }
 }