+++ lib/classes/EzWebLink.dart
@@ -4,26 +4,27 @@ import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/gestures.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class EzWebLink extends TextSpan {
-  /// Use with [EzWebLink.onTap] to setup a quick link text span
-  const EzWebLink({
+  /// Creates a [TextSpan] with an external link via [TapGestureRecognizer] && [launchUrl]
+  /// See [EzLink] for making internal links
+  ///
+  /// --- WARNING! ---
+  /// Unfortunately, the context (right-click) menu will not work as expected here
+  /// Current theory: the [TextSpan] clobbers it, but...
+  /// Empathetech chose to prioritize selectable text >> context menus
+  ///
+  /// Requires [semanticsLabel] to enforce accessibility
+  EzWebLink({
     required String text,
-    required GestureRecognizer recognizer,
+    required Uri url,
     TextStyle? style,
-    MouseCursor? mouseCursor,
-    void Function(PointerEnterEvent)? onEnter,
-    void Function(PointerExitEvent)? onExit,
+    required String semanticsLabel,
   }) : super(
           text: text,
-          recognizer: recognizer,
+          recognizer: new TapGestureRecognizer()..onTap = () => launchUrl(url),
           style: style,
-          mouseCursor: mouseCursor,
-          onEnter: onEnter,
-          onExit: onExit,
+          semanticsLabel: semanticsLabel,
         );
-
-  static GestureRecognizer onTap({required Uri url}) {
-    return TapGestureRecognizer()..onTap = () => openLink(url);
-  }
 }